+++ app/javascript/services/index.js
@@ -1,3 +1,6 @@
+import { ActivatedMenu } from '../components/component_activated_menu';
+
+
 // Always runs when document ready.
 //
 $(document).ready(function() {
@@ -44,63 +47,6 @@ function applyFormDialogs() {
       submit_text: $form.find(":submit").val()
     });
   });
-}
-
-
-class ActivatedMenu {
-  constructor($menu, config) {
-    this.activator = $("<button class=\"ActivatedMenu_Activator\"></button>");
-    this.container = $("<div class=\"ActivatedMenu_Container\"></div>");
-    this.menu = $menu;
-    this.state = {
-      open: false
-    }
-
-    this.menu.before(this.container);
-    this.menu.menu(config.menu); // Bit confusing but is how jQueryUI adds effect to eleemnt.
-
-    this.menu.on("menuselect", function() {
-      console.log("Menu select");
-    });
-
-    if(config.activator_classname) {
-      this.activator.addClass(config.activator_classname);
-    }
-
-    this.activator.text(config.activator_text);
-    this.activator.on("click.ActivatedMenu", () => {
-      if(this.state.open) {
-        this.close();
-      }
-      else {
-        this.open();
-      }
-    });
-
-    this.container.append(this.menu);
-    this.container.before(this.activator);
-
-    this.close();
-  }
-
-  // Method
-  open() {
-    console.log("ActivateMenu.open");
-    this.container.show();
-    this.state.open = true;
-  }
-
-  // Method
-  close() {
-    console.log("ActivateMenu.close");
-    this.container.hide();
-    this.state.open = false;
-  }
-
-  // Method
-  action() {
-    console.log("ActivateMenu.action");
-  }
 }
