--- app/javascript/styles/_component_activated_menu.scss
@@ -1,27 +1,0 @@
-/**
- * Custom component: Activated Menu
- * Written By: steven.burnell@digital.justice.gov.uk
- * Documentation:
- *     - activated_menu.js (fb-editor file).
- *     - https://api.jqueryui.com/menu
- *
- * ----------------------------------------------------
- *
- * fb-editor specific CSS for styling custom component.
- * This component wraps the third-party jQueryUI Menu.
- *
- **/
-
-
-.component-activated-menu {
-  a {
-    @include button_type_link;
-    border: none;
-  }
-
-  &, ul {
-    border: govuk-colour("blue") solid 2px;
-    position: absolute;
-    width: 200px;
-  }
-}
