+++ app/javascript/styles/application.scss
@@ -17,4 +17,4 @@
 @import "../packs/stylesheets/govuk";
 @import "./index";
 @import "./component_dialog";
-@import "./component_activated_menu";
+@import "../components/component_activated_menu";
